Move OptionsFiles to shared options
