+++ pkg/provider/gardener/ruleset/disak8sstig/v1r11_ruleset.go
@@ -13,10 +13,10 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/gardener/diki/pkg/config"
-	dikiutils "github.com/gardener/diki/pkg/internal/utils"
 	"github.com/gardener/diki/pkg/kubernetes/pod"
 	"github.com/gardener/diki/pkg/provider/gardener/ruleset/disak8sstig/v1r11"
 	"github.com/gardener/diki/pkg/rule"
+	shareddisastig "github.com/gardener/diki/pkg/shared/ruleset/disak8sstig"
 	sharedv1r11 "github.com/gardener/diki/pkg/shared/ruleset/disak8sstig/v1r11"
 )
 
@@ -109,7 +109,7 @@ func (r *Ruleset) registerV1R11Rules(ruleOptions map[string]config.RuleOptionsCo
 		return err
 	}
 
-	optsPodFiles, err := getV1R11OptionOrNil[dikiutils.OptionsFiles](ruleOptions[v1r11.IDPodFiles].Args)
+	optsPodFiles, err := getV1R11OptionOrNil[shareddisastig.OptionsFiles](ruleOptions[v1r11.IDPodFiles].Args)
 	if err != nil {
 		return err
 	}
